refactor: moved getStateFlips, validateMove to boardState directory
