+++ src/lib/boardState/__test__/getStateFlips.test.ts
