+++ src/lib/boardState/__test__/validateMove.test.ts
