+++ src/lib/boardState/getStateFlips.ts
